Move create_test_data to utils/
